backend/requirements.txt
asgiref             3.8.1
Django              5.2
django-cors-headers 4.7.0
djangorestframework 3.16.0
pip                 25.0.1
python-dotenv       1.1.0
pytrovich           0.0.4
sqlparse            0.5.3
tzdata              2025.2